+++ cmd/test/README.md
@@ -1,4 +1,2 @@
 test
 ====
-
-The test tests swarm of nodes and cli.
